feat: add product creation functionality to Dashboard with success message alert
