+++ src/features/sales/SalesForm.tsx
@@ -1,4 +1,4 @@
-import { useState, useEffect, useCallback, useMemo } from 'react';
+import { useState, useEffect } from 'react';
 import { useParams, useNavigate } from 'react-router-dom';
 import {
   Box,
@@ -17,7 +17,6 @@ import {
   TableContainer,
   TableHead,
   TableRow,
-  TableFooter,
   Divider,
   FormControl,
   InputLabel,
@@ -36,14 +35,14 @@ import {
   FormControlLabel,
   Switch,
   Slider,
-  Checkbox
 } from '@mui/material';
-import { Save, ArrowBack, Add, Delete, Image as ImageIcon, Construction, DeleteOutline, Search, Category } from '@mui/icons-material';
+import { Save, ArrowBack, Add, Delete, Image as ImageIcon, Construction } from '@mui/icons-material';
 import { useSale, useCreateSale, useUpdateSale } from '@hooks/useSales';
-import { useItems, useCreateItem } from '@hooks/useItems';
+import { useItems } from '@hooks/useItems';
 import { Sale, SaleItem, Item, WeightUnit, LengthUnit, AreaUnit, VolumeUnit } from '@custTypes/models';
 import { formatCurrency } from '@utils/formatters';
 import LoadingScreen from '@components/ui/LoadingScreen';
+import CreateProductDialog from '@components/inventory/CreateProductDialog';
 
 export default function SaleForm() {
   const { id } = useParams();
@@ -55,7 +54,6 @@ export default function SaleForm() {
   const { data: items = [], isLoading: itemsLoading } = useItems();
   const createSale = useCreateSale();
   const updateSale = useUpdateSale(id);
-  const createItem = useCreateItem();
 
   // Form state
   const [formData, setFormData] = useState<Partial<Sale>>({
@@ -85,57 +83,6 @@ export default function SaleForm() {
   const [selectedItemPrice, setSelectedItemPrice] = useState<string>('');
 
   const [createProductDialogOpen, setCreateProductDialogOpen] = useState(false);
-  const [newProductName, setNewProductName] = useState('');
-  const [newProductSku, setNewProductSku] = useState('');
-  const [newProductCategory, setNewProductCategory] = useState('');
-  const [newProductDescription, setNewProductDescription] = useState('');
-  const [selectedMaterials, setSelectedMaterials] = useState<Array<{
-    material: Item;
-    quantity: number;
-    totalCost: number;
-  }>>([]);
-  const [materialSearchQuery, setMaterialSearchQuery] = useState('');
-  const [materialDialogOpen, setMaterialDialogOpen] = useState(false);
-  const [newProductMarkup, setNewProductMarkup] = useState(50); // Default 50% markup
-  const [categories] = useState<string[]>(() =>
-    Array.from(new Set(items.map(item => item.category).filter(Boolean)))
-  );
-
-  const [isManualPrice, setIsManualPrice] = useState<boolean>(false);
-  const [finalPrice, setFinalPrice] = useState<number>(0);
-
-  const [selectedMaterialIds, setSelectedMaterialIds] = useState<string[]>([]);
-  const [materialQuantities, setMaterialQuantities] = useState<Record<string, string>>({});
-
-  const calculateMaterialsCost = useCallback(() => {
-    return selectedMaterials.reduce((sum, { totalCost }) => sum + totalCost, 0);
-  }, [selectedMaterials]);
-
-  const calculateProductPrice = useCallback(() => {
-    const materialsCost = calculateMaterialsCost();
-    return isManualPrice ? finalPrice : materialsCost * (1 + newProductMarkup / 100);
-  }, [calculateMaterialsCost, newProductMarkup, isManualPrice, finalPrice]);
-
-  const generateProductSku = useCallback(() => {
-    const prefix = 'PROD';
-    const timestamp = Date.now().toString().slice(-6);
-    return `${prefix}${timestamp}`;
-  }, []);
-
-  useEffect(() => {
-    if (createProductDialogOpen) {
-      setNewProductSku(generateProductSku());
-    }
-  }, [createProductDialogOpen, generateProductSku]);
-
-  const filteredMaterials = useMemo(() => {
-    return items
-      .filter(item => item.itemType === 'material' || item.itemType === 'both')
-      .filter(item =>
-        item.name.toLowerCase().includes(materialSearchQuery.toLowerCase()) ||
-        item.sku.toLowerCase().includes(materialSearchQuery.toLowerCase())
-      );
-  }, [items, materialSearchQuery]);
 
   // Load existing sale data if in edit mode
   useEffect(() => {
@@ -192,16 +139,6 @@ export default function SaleForm() {
       total
     }));
   }, [formData.items, formData.taxRate, formData.discountAmount]);
-
-  useEffect(() => {
-    if (isManualPrice) {
-      // Keep current price if already set
-      return;
-    }
-    // Otherwise set default price based on materials cost and markup
-    const materialsCost = calculateMaterialsCost();
-    setFinalPrice(materialsCost * (1 + newProductMarkup / 100));
-  }, [calculateMaterialsCost, newProductMarkup, selectedMaterials, isManualPrice]);
 
   const handleInputChange = (e: React.ChangeEvent<HTMLInputElement | HTMLTextAreaElement>) => {
     const { name, value } = e.target;
@@ -477,100 +414,29 @@ export default function SaleForm() {
     return <LoadingScreen />;
   }
 
-  // Remove a material from the new product
-  const handleRemoveMaterial = (index: number) => {
-    setSelectedMaterials(prev => prev.filter((_, i) => i !== index));
-  };
+  // Handle product created from dialog
+  const handleProductCreated = (product: Item) => {
+    const productPrice = product.price;
 
-  // Create the new product and add to sale
-  const handleCreateProduct = async () => {
-    if (!newProductName) {
-      setError("Product name is required");
-      return;
-    }
+    // Add the new product to the sale
+    const newSaleItem: SaleItem = {
+      item: product._id || '',
+      name: product.name,
+      quantity: 1,
+      weight: 0,
+      length: 0,
+      area: 0,
+      volume: 0,
+      priceAtSale: productPrice,
+      soldBy: 'quantity'
+    };
 
-    if (selectedMaterials.length === 0) {
-      setError("At least one material is required");
-      return;
-    }
+    setFormData(prev => ({
+      ...prev,
+      items: [...(prev.items || []), newSaleItem]
+    }));
 
-    try {
-      // Create new product
-      const materialsCost = calculateMaterialsCost();
-      const productPrice = calculateProductPrice();
-
-      const newProduct: Partial<Item> = {
-        name: newProductName,
-        sku: newProductSku,
-        category: newProductCategory,
-        description: newProductDescription,
-        trackingType: 'quantity',
-        sellByMeasurement: 'quantity', // Add this missing property
-        quantity: 1, // Start with one item
-        weight: 0,
-        weightUnit: 'lb',
-        price: productPrice, // This now handles both pricing modes
-        priceType: 'each',
-        itemType: 'product',
-        cost: materialsCost,
-        components: selectedMaterials.map(({ material, quantity }) => ({
-          item: material._id || '',
-          quantity
-        }))
-      };
-
-      // Create the product via API
-      const createdProduct = await createItem.mutateAsync(newProduct as Item);
-
-      // Add the new product to the sale
-      const newSaleItem: SaleItem = {
-        item: createdProduct._id || '',
-        name: createdProduct.name,  // Add name for display
-        quantity: 1,
-        weight: 0,
-        length: 0,
-        area: 0,
-        volume: 0,
-        priceAtSale: productPrice,
-        soldBy: 'quantity'
-      };
-
-      setFormData(prev => ({
-        ...prev,
-        items: [...(prev.items || []), newSaleItem]
-      }));
-
-      // Reset and close dialog
-      setNewProductName('');
-      setNewProductSku('');
-      setNewProductCategory('');
-      setNewProductDescription('');
-      setSelectedMaterials([]);
-      setNewProductMarkup(50);
-      setCreateProductDialogOpen(false);
-      setError(null);
-    } catch (err) {
-      console.error('Failed to create product:', err);
-      setError('Failed to create product. Please try again.');
-    }
-  };
-
-  const handleAddMultipleMaterials = () => {
-    const newMaterials = selectedMaterialIds.map(id => {
-      const material = items.find(item => item._id === id);
-      if (!material) return null;
-
-      const quantity = parseInt(materialQuantities[id]) || 1;
-      const unitCost = material.cost || material.price;
-      const totalCost = unitCost * quantity;
-
-      return { material, quantity, totalCost };
-    }).filter(Boolean) as Array<{ material: Item; quantity: number; totalCost: number }>;
-
-    setSelectedMaterials(prev => [...prev, ...newMaterials]);
-    setSelectedMaterialIds([]);
-    setMaterialQuantities({});
-    setMaterialDialogOpen(false);
+    setError(null);
   };
 
   return (
@@ -600,7 +466,7 @@ export default function SaleForm() {
 
       {/* Form content (identical to original) */}
       <Grid2 container spacing={3}>
-        <Grid2 size= {{ xs: 12, md: 8 }}>
+        <Grid2 size={{ xs: 12, md: 8 }}>
           <Paper sx={{ p: 3, mb: 3 }}>
             <Typography variant="h6" gutterBottom>
               Customer Information
@@ -608,7 +474,7 @@ export default function SaleForm() {
             <Divider sx={{ mb: 2 }} />
 
             <Grid2 container spacing={2}>
-              <Grid2 size= {{ xs: 12, sm: 12 }}>
+              <Grid2 size={{ xs: 12, sm: 12 }}>
                 <TextField
                   fullWidth
                   label="Customer Name"
@@ -619,7 +485,7 @@ export default function SaleForm() {
                   placeholder="Optional"
                 />
               </Grid2>
-              <Grid2 size= {{ xs: 12, sm: 6 }}>
+              <Grid2 size={{ xs: 12, sm: 6 }}>
                 <TextField
                   fullWidth
                   label="Email"
@@ -630,7 +496,7 @@ export default function SaleForm() {
                   placeholder="Optional"
                 />
               </Grid2>
-              <Grid2 size= {{ xs: 12, sm: 6 }}>
+              <Grid2 size={{ xs: 12, sm: 6 }}>
                 <TextField
                   fullWidth
                   label="Phone"
@@ -701,14 +567,14 @@ export default function SaleForm() {
                             item.soldBy === 'quantity'
                               ? item.priceAtSale * item.quantity
                               : item.soldBy === 'weight'
-                              ? item.priceAtSale * item.weight
-                              : item.soldBy === 'length'
-                              ? item.priceAtSale * item.length
-                              : item.soldBy === 'area'
-                              ? item.priceAtSale * item.area
-                              : item.soldBy === 'volume'
-                              ? item.priceAtSale * item.volume
-                              : item.priceAtSale
+                                ? item.priceAtSale * item.weight
+                                : item.soldBy === 'length'
+                                  ? item.priceAtSale * item.length
+                                  : item.soldBy === 'area'
+                                    ? item.priceAtSale * item.area
+                                    : item.soldBy === 'volume'
+                                      ? item.priceAtSale * item.volume
+                                      : item.priceAtSale
                           )}
                         </TableCell>
                         <TableCell align="right">
@@ -741,7 +607,7 @@ export default function SaleForm() {
             <Divider sx={{ mb: 2 }} />
 
             <Grid2 container spacing={2}>
-              <Grid2 size= {{ xs: 12, sm: 6 }}>
+              <Grid2 size={{ xs: 12, sm: 6 }}>
                 <FormControl fullWidth>
                   <InputLabel>Payment Method</InputLabel>
                   <Select
@@ -759,7 +625,7 @@ export default function SaleForm() {
                   </Select>
                 </FormControl>
               </Grid2>
-              <Grid2 size= {{ xs: 12, sm: 6 }}>
+              <Grid2 size={{ xs: 12, sm: 6 }}>
                 <FormControl fullWidth>
                   <InputLabel>Status</InputLabel>
                   <Select
@@ -775,7 +641,7 @@ export default function SaleForm() {
                   </Select>
                 </FormControl>
               </Grid2>
-              <Grid2 size= {{ xs: 12 }}>
+              <Grid2 size={{ xs: 12 }}>
                 <TextField
                   fullWidth
                   multiline
@@ -792,7 +658,7 @@ export default function SaleForm() {
           </Paper>
         </Grid2>
 
-        <Grid2 size= {{ xs: 12, md: 4 }}>
+        <Grid2 size={{ xs: 12, md: 4 }}>
           <Paper sx={{ p: 3, position: 'sticky', top: 20 }}>
             <Typography variant="h6" gutterBottom>
               Sale Summary
@@ -801,16 +667,16 @@ export default function SaleForm() {
 
             <Box sx={{ mb: 2 }}>
               <Grid2 container spacing={2}>
-                <Grid2 size= {{ xs: 7 }}>
+                <Grid2 size={{ xs: 7 }}>
                   <Typography variant="body1">Subtotal:</Typography>
                 </Grid2>
-                <Grid2 size= {{ xs: 5 }}>
+                <Grid2 size={{ xs: 5 }}>
                   <Typography variant="body1" align="right">
                     {formatCurrency(formData.subtotal || 0)}
                   </Typography>
                 </Grid2>
 
-                <Grid2 size= {{ xs: 7 }} sx={{ display: 'flex', alignItems: 'center' }}>
+                <Grid2 size={{ xs: 7 }} sx={{ display: 'flex', alignItems: 'center' }}>
                   <Typography variant="body1" sx={{ mr: 1 }}>Tax Rate:</Typography>
                   <TextField
                     size="small"
@@ -826,13 +692,13 @@ export default function SaleForm() {
                     sx={{ width: 80 }}
                   />
                 </Grid2>
-                <Grid2 size= {{ xs: 5 }}>
+                <Grid2 size={{ xs: 5 }}>
                   <Typography variant="body1" align="right">
                     {formatCurrency(formData.taxAmount || 0)}
                   </Typography>
                 </Grid2>
 
-                <Grid2 size= {{ xs: 7 }} sx={{ display: 'flex', alignItems: 'center' }}>
+                <Grid2 size={{ xs: 7 }} sx={{ display: 'flex', alignItems: 'center' }}>
                   <Typography variant="body1" sx={{ mr: 1 }}>Discount:</Typography>
                   <TextField
                     size="small"
@@ -848,7 +714,7 @@ export default function SaleForm() {
                     sx={{ width: 100 }}
                   />
                 </Grid2>
-                <Grid2 size= {{ xs: 5 }}>
+                <Grid2 size={{ xs: 5 }}>
                   <Typography variant="body1" align="right" color="error">
                     -{formatCurrency(formData.discountAmount || 0)}
                   </Typography>
@@ -860,10 +726,10 @@ export default function SaleForm() {
 
             <Box sx={{ mb: 3 }}>
               <Grid2 container spacing={1}>
-                <Grid2 size= {{ xs: 6 }}>
+                <Grid2 size={{ xs: 6 }}>
                   <Typography variant="h6">Total:</Typography>
                 </Grid2>
-                <Grid2 size= {{ xs: 6 }}>
+                <Grid2 size={{ xs: 6 }}>
                   <Typography variant="h6" align="right" color="primary">
                     {formatCurrency(formData.total || 0)}
                   </Typography>
@@ -896,25 +762,25 @@ export default function SaleForm() {
             <Box sx={{ mb: 2 }}>
               <Typography variant="h6" gutterBottom>{selectedItem.name}</Typography>
               <Grid2 container spacing={2}>
-                <Grid2 size= {{ xs: 12 }}>
+                <Grid2 size={{ xs: 12 }}>
                   <Typography variant="body2" color="text.secondary">
                     SKU: {selectedItem.sku} | Category: {selectedItem.category}
                   </Typography>
                 </Grid2>
-                <Grid2 size= {{ xs: 12 }}>
+                <Grid2 size={{ xs: 12 }}>
                   <Typography variant="body1">
                     {selectedItem.trackingType === 'quantity'
                       ? `Available: ${selectedItem.quantity} in stock`
                       : `Available: ${selectedItem.weight} ${selectedItem.weightUnit} in stock`}
                   </Typography>
                 </Grid2>
-                <Grid2 size= {{ xs: 12 }}>
+                <Grid2 size={{ xs: 12 }}>
                   <Typography variant="body1">
                     Price: {formatCurrency(selectedItem.price)}
                     {selectedItem.priceType === 'per_weight_unit' && `/${selectedItem.weightUnit}`}
                   </Typography>
                 </Grid2>
-                <Grid2 size= {{ xs: 12 }}>
+                <Grid2 size={{ xs: 12 }}>
                   {selectedItem.trackingType === 'quantity' ? (
                     <TextField
                       fullWidth
@@ -1211,373 +1077,13 @@ export default function SaleForm() {
         </DialogActions>
       </Dialog>
 
-      {/* Create Product Dialog */}
-      <Dialog
+      {/* Use the new CreateProductDialog component */}
+      <CreateProductDialog
         open={createProductDialogOpen}
         onClose={() => setCreateProductDialogOpen(false)}
-        fullWidth
-        maxWidth="md"
-      >
-        <DialogTitle>Create New Product</DialogTitle>
-        <DialogContent dividers>
-          <Grid2 container spacing={3}>
-            <Grid2 size={{ xs: 12, sm: 6 }}>
-              <TextField
-                fullWidth
-                required
-                label="Product Name"
-                value={newProductName}
-                onChange={(e) => setNewProductName(e.target.value)}
-                margin="normal"
-              />
-            </Grid2>
-            <Grid2 size={{ xs: 12, sm: 6 }}>
-              <TextField
-                fullWidth
-                required
-                label="SKU"
-                value={newProductSku}
-                onChange={(e) => setNewProductSku(e.target.value)}
-                margin="normal"
-              />
-            </Grid2>
-            <Grid2 size={{ xs: 12, sm: 6 }}>
-              <FormControl fullWidth margin="normal">
-                <InputLabel>Category</InputLabel>
-                <Select
-                  value={newProductCategory}
-                  onChange={(e) => setNewProductCategory(e.target.value)}
-                  label="Category"
-                >
-                  <MenuItem value="">None</MenuItem>
-                    {(categories as string[]).map((category) => (
-                    <MenuItem key={category} value={category}>
-                      {category}
-                    </MenuItem>
-                    ))}
-                </Select>
-              </FormControl>
-            </Grid2>
-            <Grid2 size={{ xs: 12, sm: 6 }}>
-              <Box sx={{ mt: 2, mb: 1 }}>
-                <Typography variant="subtitle2">Materials Cost</Typography>
-                <Typography variant="h6" color="text.secondary">
-                  {formatCurrency(calculateMaterialsCost())}
-                </Typography>
-              </Box>
-            </Grid2>
-            <Grid2 size={{ xs: 12 }}>
-              <TextField
-                fullWidth
-                multiline
-                rows={2}
-                label="Description"
-                value={newProductDescription}
-                onChange={(e) => setNewProductDescription(e.target.value)}
-                margin="normal"
-              />
-            </Grid2>
-          </Grid2>
+        onProductCreated={handleProductCreated}
+      />
 
-          <Divider sx={{ my: 3 }} />
-          <Typography variant="h6" gutterBottom>Materials</Typography>
-
-          <Box sx={{ mb: 2 }}>
-            <Button
-              variant="outlined"
-              startIcon={<Add />}
-              onClick={() => setMaterialDialogOpen(true)}
-            >
-              Add Material
-            </Button>
-          </Box>
-
-          {selectedMaterials.length > 0 ? (
-            <TableContainer component={Paper} variant="outlined">
-              <Table size="small">
-                <TableHead>
-                  <TableRow>
-                    <TableCell>Material</TableCell>
-                    <TableCell align="right">Cost</TableCell>
-                    <TableCell align="center">Quantity</TableCell>
-                    <TableCell align="right">Total</TableCell>
-                    <TableCell></TableCell>
-                  </TableRow>
-                </TableHead>
-                <TableBody>
-                  {selectedMaterials.map((item, index) => (
-                    <TableRow key={index}>
-                      <TableCell>{item.material.name}</TableCell>
-                      <TableCell align="right">{formatCurrency(item.material.cost || item.material.price)}</TableCell>
-                      <TableCell align="center">{item.quantity}</TableCell>
-                      <TableCell align="right">{formatCurrency(item.totalCost)}</TableCell>
-                      <TableCell>
-                        <IconButton
-                          size="small"
-                          color="error"
-                          onClick={() => handleRemoveMaterial(index)}
-                        >
-                          <DeleteOutline fontSize="small" />
-                        </IconButton>
-                      </TableCell>
-                    </TableRow>
-                  ))}
-                </TableBody>
-                <TableFooter>
-                  <TableRow>
-                    <TableCell colSpan={3} align="right">
-                      <Typography variant="subtitle2">Total Materials Cost:</Typography>
-                    </TableCell>
-                    <TableCell align="right" colSpan={2}>
-                      <Typography variant="subtitle1" fontWeight="bold">
-                        {formatCurrency(calculateMaterialsCost())}
-                      </Typography>
-                    </TableCell>
-                  </TableRow>
-                </TableFooter>
-              </Table>
-            </TableContainer>
-          ) : (
-            <Alert severity="info" sx={{ mb: 3 }}>
-              No materials added yet. Click "Add Material" to add materials to this product.
-            </Alert>
-          )}
-
-          <Divider sx={{ my: 3 }} />
-
-          <Typography variant="h6" gutterBottom>Pricing</Typography>
-          <Grid2 container spacing={2} alignItems="center">
-            <Grid2 size={{ xs: 12 }}>
-              <FormControlLabel
-                control={
-                  <Switch
-                    checked={!isManualPrice}
-                    onChange={() => setIsManualPrice(!isManualPrice)}
-                  />
-                }
-                label={isManualPrice ? "Set Final Price" : "Set Markup Percentage"}
-              />
-            </Grid2>
-
-            {!isManualPrice ? (
-              // Markup Percentage Mode
-              <Grid2 size={{ xs: 12, md: 6 }}>
-                <Box>
-                  <Typography variant="body2" gutterBottom>
-                    Markup percentage: {newProductMarkup}%
-                  </Typography>
-                  <Slider
-                    value={newProductMarkup}
-                    onChange={(_, newValue) => setNewProductMarkup(newValue as number)}
-                    step={5}
-                    marks
-                    min={0}
-                    max={1000}
-                    valueLabelDisplay="auto"
-                    valueLabelFormat={(value) => `${value}%`}
-                  />
-                </Box>
-              </Grid2>
-            ) : (
-              // Final Price Mode
-              <Grid2 size={{ xs: 12, md: 6 }}>
-                <TextField
-                  fullWidth
-                  label="Final Price"
-                  type="number"
-                  value={finalPrice}
-                  onChange={(e) => {
-                    const price = parseFloat(e.target.value) || 0;
-                    setFinalPrice(price);
-                    // Calculate markup percentage based on materials cost
-                    const materialsCost = calculateMaterialsCost();
-                    if (materialsCost > 0) {
-                      const calculatedMarkup = ((price / materialsCost) - 1) * 100;
-                      setNewProductMarkup(Math.round(calculatedMarkup));
-                    }
-                  }}
-                  InputProps={{
-                    startAdornment: <InputAdornment position="start">$</InputAdornment>,
-                    inputProps: { min: 0, step: 0.01 }
-                  }}
-                />
-              </Grid2>
-            )}
-
-            <Grid2 size={{ xs: 12, md: 6 }}>
-              <Paper variant="outlined" sx={{ p: 2, bgcolor: 'background.default' }}>
-                <Typography variant="body2" color="text.secondary">
-                  Materials Cost: {formatCurrency(calculateMaterialsCost())}
-                </Typography>
-                <Typography variant="body2" color="success.main">
-                  Markup: {formatCurrency(calculateProductPrice() - calculateMaterialsCost())}
-                </Typography>
-                <Divider sx={{ my: 1 }} />
-                <Typography variant="h6" fontWeight="bold" color="primary">
-                  Final Price: {formatCurrency(calculateProductPrice())}
-                </Typography>
-              </Paper>
-            </Grid2>
-          </Grid2>
-        </DialogContent>
-        <DialogActions>
-          <Button onClick={() => setCreateProductDialogOpen(false)}>Cancel</Button>
-          <Button
-            onClick={handleCreateProduct}
-            variant="contained"
-            color="primary"
-            disabled={!newProductName || selectedMaterials.length === 0}
-          >
-            Create & Add to Sale
-          </Button>
-        </DialogActions>
-      </Dialog>
-
-      {/* Material Selection Dialog */}
-      <Dialog
-        open={materialDialogOpen}
-        onClose={() => setMaterialDialogOpen(false)}
-        fullWidth
-        maxWidth="md" // Changed from "sm" to "md" for more space
-      >
-        <DialogTitle>
-          Select Materials
-          <Typography variant="subtitle2" color="text.secondary">
-            Select multiple materials to add at once
-          </Typography>
-        </DialogTitle>
-        <DialogContent dividers>
-          <TextField
-            fullWidth
-            label="Search Materials"
-            value={materialSearchQuery}
-            onChange={(e) => setMaterialSearchQuery(e.target.value)}
-            margin="normal"
-            InputProps={{
-              startAdornment: (
-                <InputAdornment position="start">
-                  <Search />
-                </InputAdornment>
-              ),
-            }}
-          />
-
-          <TableContainer component={Paper} variant="outlined" sx={{ mt: 2 }}>
-            <Table size="small">
-              <TableHead>
-                <TableRow>
-                  <TableCell padding="checkbox">
-                    <Checkbox
-                      indeterminate={selectedMaterialIds.length > 0 && selectedMaterialIds.length < filteredMaterials.length}
-                      checked={filteredMaterials.length > 0 && selectedMaterialIds.length === filteredMaterials.length}
-                      onChange={(e) => {
-                        if (e.target.checked) {
-                          // Select all materials
-                          setSelectedMaterialIds(filteredMaterials.map(m => m._id || ''));
-                          // Initialize all quantities to 1
-                          const newQuantities: Record<string, string> = {};
-                          filteredMaterials.forEach(m => {
-                            if (m._id) newQuantities[m._id] = '1';
-                          });
-                          setMaterialQuantities({...materialQuantities, ...newQuantities});
-                        } else {
-                          // Unselect all
-                          setSelectedMaterialIds([]);
-                        }
-                      }}
-                    />
-                  </TableCell>
-                  <TableCell>Material</TableCell>
-                  <TableCell align="right">Cost</TableCell>
-                  <TableCell align="center">In Stock</TableCell>
-                  <TableCell align="right">Quantity</TableCell>
-                </TableRow>
-              </TableHead>
-              <TableBody>
-                {filteredMaterials.map(item => (
-                  <TableRow
-                    key={item._id}
-                    selected={item._id ? selectedMaterialIds.includes(item._id) : false}
-                    sx={{
-                      '&.Mui-selected': {
-                        backgroundColor: 'rgba(25, 118, 210, 0.08)'
-                      }
-                    }}
-                  >
-                    <TableCell padding="checkbox">
-                      <Checkbox
-                        checked={item._id ? selectedMaterialIds.includes(item._id) : false}
-                        onChange={(e) => {
-                          if (e.target.checked) {
-                            // Add to selection
-                            setSelectedMaterialIds(prev => [...prev, item._id || '']);
-                            // Set default quantity
-                            setMaterialQuantities(prev => ({
-                              ...prev,
-                              [item._id || '']: '1'
-                            }));
-                          } else {
-                            // Remove from selection
-                            setSelectedMaterialIds(prev => prev.filter(id => id !== item._id));
-                          }
-                        }}
-                      />
-                    </TableCell>
-                    <TableCell>
-                      <Box sx={{ display: 'flex', alignItems: 'center' }}>
-                        <Avatar
-                          src={item.imageUrl || undefined}
-                          variant="rounded"
-                          sx={{ width: 40, height: 40, mr: 2 }}
-                        >
-                          <Category fontSize="small" />
-                        </Avatar>
-                        <Box>
-                          <Typography variant="body1">{item.name}</Typography>
-                          <Typography variant="body2" color="text.secondary">SKU: {item.sku}</Typography>
-                        </Box>
-                      </Box>
-                    </TableCell>
-                    <TableCell align="right">{formatCurrency(item.cost || item.price)}</TableCell>
-                    <TableCell align="center">
-                      {item.trackingType === 'quantity'
-                        ? `${item.quantity} units`
-                        : `${item.weight} ${item.weightUnit}`}
-                    </TableCell>
-                    <TableCell align="right">
-                      {item._id && selectedMaterialIds.includes(item._id) && (
-                        <TextField
-                          type="number"
-                          size="small"
-                          value={materialQuantities[item._id] || '1'}
-                          onChange={(e) => setMaterialQuantities(prev => ({
-                            ...prev,
-                            [item._id as string]: e.target.value  // Add "as string" here
-                          }))}
-                          InputProps={{ inputProps: { min: 1 } }}
-                          sx={{ width: 70 }}
-                          onClick={(e) => e.stopPropagation()}
-                        />
-                      )}
-                    </TableCell>
-                  </TableRow>
-                ))}
-              </TableBody>
-            </Table>
-          </TableContainer>
-        </DialogContent>
-        <DialogActions>
-          <Button onClick={() => setMaterialDialogOpen(false)}>Cancel</Button>
-          <Button
-            onClick={handleAddMultipleMaterials}
-            variant="contained"
-            color="primary"
-            disabled={selectedMaterialIds.length === 0}
-          >
-            Add Selected Materials
-          </Button>
-        </DialogActions>
-      </Dialog>
     </Box>
   );
 }
